relocate functions out of cli
