+++ src/filter/index.ts
@@ -1,4 +1,5 @@
 import { filterPlugin } from './filter_plugin'
 export * from './types'
+export { orderPickles } from './order_pickles'
 
 export default filterPlugin
